requirements.txt
Pillow>=10.0.0
rembg>=2.0.50
numpy>=1.24.0
onnxruntime>=1.16.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.0
